Reorganized Flask routes and template files
