+++ scoring/web/templates/auth/login.html
@@ -9,7 +9,7 @@
             <strong>{{ error }}</strong>
         </div>
     {% endif %}
-    <form method="POST" action="{{ url_for('login') }}">
+    <form method="POST" action="{{ url_for('auth.login') }}">
         {{ form.csrf_token }}
         <div class="form-group">
             {{ form.username.label }}
