--- scoring/web/templates/competition.html
@@ -1,23 +1,0 @@
-{% extends 'base.html' %}
-
-{% block settings %}active{% endblock %}
-{% block content %}
-<h1>Competition Controls</h1>
-{% if running %}
-<h2>Status: <strong class="text-success">Running</strong></h2>
-{% else %}
-<h2>Status: <strong class="text-danger">Stopped</strong></h2>
-{% endif %}
-
-<form method="POST" action="{{ url_for('settings') }}">
-    <div class="form-group">
-    <label for="running">Start/Stop: </label>
-    {% if running %}
-    <input type="submit" class="btn btn-danger" name="running" value="Stop"/>
-    {% else %}
-    <input type="submit" class="btn btn-success" name="running" value="Start"/>
-    {% endif %}
-    </div>
-</form>
-
-{% endblock %}
