--- scoring/web/templates/index.html
@@ -1,29 +1,0 @@
-{% extends 'base.html' %}
-
-{% block index %}active{% endblock %}
-
-{% block head %}<meta http-equiv="refresh" content="5">{% endblock %}
-
-{% block content %}
-<h1>Scoring Engine</h1>
-
-<div class="col-md-6">
-<div class="panel panel-default">
-<div class="panel-body">
-<table class="table">
-<thead>
-<tr>
-<th>Service</th>
-<th>Status</th>
-</tr>
-</thead>
-
-<tbody>
-</tbody>
-</table>
-</div>
-</div>
-</a>
-</div>
-
-{% endblock %}
