+++ scoring/web/templates/pcr/pcr_overview.html
@@ -18,7 +18,7 @@
         <th>Domain/Service</th>
         <th>Status</th>
         <th>
-            <a href="{{ url_for('new_pcr') }}" class="btn btn-success">
+            <a href="{{ url_for('pcr.new_pcr') }}" class="btn btn-success">
                 New Request
             </a>
         </th>
@@ -36,7 +36,7 @@
             </td>
             <td>
             {% if pcr.check_id is none %}
-                {{ domains[pcr.domain_id] }}
+                {{ pcr.domain }}
             {% else %}
                 {% set c = checks[pcr.check_id] %}
                 {{ c.name }}
@@ -52,7 +52,7 @@
             <td class="text-danger"><strong>Denied</strong></td>
             {% endif %}
             <td>
-                <a href="{{ url_for('pcr_details') }}?id={{pcr.id}}" class="btn btn-primary">
+                <a href="{{ url_for('pcr.pcr_details') }}?id={{pcr.id}}" class="btn btn-primary">
                     View
                 </a>
                 {% if not current_user.is_admin and pcr.status != 0 %}
@@ -79,7 +79,7 @@
         Are you sure you want to delete this password change request?
       </div>
       <div class="modal-footer">
-        <form method="post" action="{{ url_for('pcr') }}" id="delete-pcr">
+        <form method="post" action="{{ url_for('pcr.pcr') }}" id="delete-pcr">
             <input type="hidden" id="reqId" name="reqId" value="0"/>
             <button type="submit" class="btn btn-danger">Delete</button>
         </form>
