+++ scoring/web/templates/reports/score.html
@@ -6,7 +6,7 @@
 <script type="text/javascript" src="/static/js/Chart.bundle-2.7.2.min.js"></script>
 <script src="/static/js/score.js"></script>
 
-<form action="{{ url_for('score') }}" method="GET" style="margin-top: 20px; margin-left:12.5%;">
+<form action="{{ url_for('reporting.score') }}" method="GET" style="margin-top: 20px; margin-left:12.5%;">
 <label for="start">Start Time:</label>
 <input type="time" id="start-time" name="start"/>
