+++ scoring/web/templates/result_log.html
@@ -3,8 +3,6 @@
 {% block content %}
 <h1>Result Log</h1>
 
-<img src={{ url_for('static', filename='imgs/'+fname) }}/>
-
 <table class="table">
 <thead>
 <tr>
