+++ scoring/web/templates/vcloud/revert_log.html
