+++ scoring/wsgi.py
@@ -1,4 +1,4 @@
-from web.server import app as application
+from web.app import app as application
 
 if __name__ == '__main__':
     application.run()
